+++ index.js
@@ -1,44 +1,27 @@
 "use strict";
 
-const path = require("path");
-const fs = require("fs");
-const getRules = require("./rules");
+const reactPlugin = require("eslint-plugin-react");
+const reactRules = require("./react-rules");
+const baseConfig = require("./base-config");
+const testConfig = require("./test");
 
-// This will take care of potential symlinks
-const appDir = fs.realpathSync(process.cwd());
-
-const isModuleProject = require(path.resolve(appDir, "package.json")).type === "module";
-const hasES2022Support = parseInt(process.versions.node.split(".").shift(), 10) >= 16;
-
-const moduleConfig = {
-  parserOptions: {
-    ecmaVersion: hasES2022Support ? 2022 : 2021,
-    sourceType: "module",
+module.exports = [
+  baseConfig,
+  {
+    ...baseConfig,
+    files: [ "**/*.jsx" ],
+    languageOptions: {
+      ...baseConfig.languageOptions,
+      parserOptions: { ecmaFeatures: { jsx: true } },
+    },
+    plugins: { ...baseConfig.plugins, react: reactPlugin },
+    rules: { ...baseConfig.rules, ...reactRules },
   },
-  env: {
-    node: true,
-    ...(hasES2022Support ? { es2022: true } : { es6: true }),
+  {
+    ...baseConfig,
+    languageOptions: { ...baseConfig.languageOptions, ...testConfig.languageOptions },
+    plugins: { ...baseConfig.plugins, ...testConfig.plugins },
+    rules: { ...baseConfig.rules, ...testConfig.rules },
+    files: [ "test/**/*.js" ],
   },
-  plugins: [ "eslint-plugin-n", "import", "@bonniernews/typescript-rules" ],
-};
-
-const commonjsConfig = {
-  parserOptions: { ecmaVersion: 2021 },
-  env: {
-    node: true,
-    es6: true,
-  },
-  plugins: [ "eslint-plugin-n", "@bonniernews/typescript-rules" ],
-};
-
-module.exports = {
-  ignorePatterns: [
-    "tmp/",
-    "public/",
-    "submodule/**",
-    "logs/",
-    "docs/",
-  ],
-  ...(isModuleProject ? moduleConfig : commonjsConfig),
-  rules: getRules(isModuleProject),
-};
+];